requirements.txt
pillow>=11.2.1
PyYAML>=6.0
tqdm>=4.64.0 